# speed up size, lbounds, and ubounds with @generated functions