`size()`, `lbounds()`, and `ubounds()` currently use the `ntuple(i -> f(i), N)` syntax, which is rather slow. A generated function makes the return type inferable without any type assertions and also improves performance. It also has the side-effect of fixing an apparent typeassert bug in `lbounds` and `ubounds`. 

## Before

```julia
julia> using BenchmarkTools

julia> A = rand(2,2,2)
2×2×2 Array{Float64,3}:
[:, :, 1] =
 0.771663  0.124192
 0.524341  0.676021

[:, :, 2] =
 0.295365  0.966051
 0.585499  0.599789

julia> itp = interpolate(A, BSpline(Linear()), OnGrid())
2×2×2 Interpolations.BSplineInterpolation{Float64,3,Array{Float64,3},Interpolations.BSpline{Interpolations.Linear},Interpolations.OnGrid,0}:
[:, :, 1] =
 0.771663  0.124192
 0.524341  0.676021

[:, :, 2] =
 0.295365  0.966051
 0.585499  0.599789

julia> @benchmark size($itp)
BenchmarkTools.Trial:
  memory estimate:  48.00 bytes
  allocs estimate:  2
  --------------
  minimum time:     75.793 ns (0.00% GC)
  median time:      78.353 ns (0.00% GC)
  mean time:        90.578 ns (3.68% GC)
  maximum time:     3.053 μs (93.58% GC)
  --------------
  samples:          10000
  evals/sample:     970
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Interpolations.lbounds($itp)
ERROR: TypeError: typeassert: expected Tuple{Float64,Float64,Float64}, got Tuple{Int64,Int64,Int64}
 in ##core#280() at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:250
 in ##sample#281(::BenchmarkTools.Parameters) at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:257
 in #_run#2(::Bool, ::String, ::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#279")}, ::BenchmarkTools.Parameters) at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:285
 in (::BenchmarkTools.#kw##_run)(::Array{Any,1}, ::BenchmarkTools.#_run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#279")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in anonymous at ./<missing>:?
 in #run#17(::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#279")}, ::BenchmarkTools.Parameters) at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:32
 in (::Base.#kw##run)(::Array{Any,1}, ::Base.#run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#279")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in warmup(::BenchmarkTools.Benchmark{Symbol("##benchmark#279")}) at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:67

julia> @benchmark Interpolations.ubounds($itp)
ERROR: TypeError: typeassert: expected Tuple{Float64,Float64,Float64}, got Tuple{Int64,Int64,Int64}
 in ##core#284() at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:250
 in ##sample#285(::BenchmarkTools.Parameters) at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:257
 in #_run#3(::Bool, ::String, ::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#283")}, ::BenchmarkTools.Parameters) at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:285
 in (::BenchmarkTools.#kw##_run)(::Array{Any,1}, ::BenchmarkTools.#_run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#283")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in anonymous at ./<missing>:?
 in #run#17(::Array{Any,1}, ::Function, ::BenchmarkTools.Benchmark{Symbol("##benchmark#283")}, ::BenchmarkTools.Parameters) at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:32
 in (::Base.#kw##run)(::Array{Any,1}, ::Base.#run, ::BenchmarkTools.Benchmark{Symbol("##benchmark#283")}, ::BenchmarkTools.Parameters) at ./<missing>:0
 in warmup(::BenchmarkTools.Benchmark{Symbol("##benchmark#283")}) at /Users/rdeits/.julia/v0.5/BenchmarkTools/src/execution.jl:67
```

## After

```julia
julia> using BenchmarkTools

julia> A = rand(2,2,2);

julia> itp = interpolate(A, BSpline(Linear()), OnGrid());

julia> @benchmark size($itp)
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     3.691 ns (0.00% GC)
  median time:      3.752 ns (0.00% GC)
  mean time:        3.836 ns (0.00% GC)
  maximum time:     16.309 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Interpolations.lbounds($itp)
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     1.695 ns (0.00% GC)
  median time:      1.739 ns (0.00% GC)
  mean time:        1.785 ns (0.00% GC)
  maximum time:     24.235 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Interpolations.ubounds($itp)
BenchmarkTools.Trial:
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     5.964 ns (0.00% GC)
  median time:      5.989 ns (0.00% GC)
  mean time:        6.923 ns (0.00% GC)
  maximum time:     41.035 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%
```