@rdeits 2016-12-04
Also, if you'd rather avoid `@generated` functions here, a trivial improvement would be to change `ntuple(i -> f(i), N)` to `ntuple(i -> f(i), Val{N})`.