@timholy 2017-02-13
Sorry I didn't notice this. Yes, it would be better to use the `Val{N}` form. Would you mind?