@rdeits 2017-02-13
Ok, I've switched to the `Val{N}` form and rebased. Performance is slightly worse than `@generated` but still an improvement over the original:

```julia
julia> @benchmark size($itp)
BenchmarkTools.Trial:
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     17.439 ns (0.00% GC)
  median time:      19.440 ns (0.00% GC)
  mean time:        22.403 ns (3.33% GC)
  maximum time:     1.231 μs (96.20% GC)
  --------------
  samples:          10000
  evals/sample:     998
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Interpolations.lbounds($itp)
BenchmarkTools.Trial:
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     11.905 ns (0.00% GC)
  median time:      13.343 ns (0.00% GC)
  mean time:        15.039 ns (5.08% GC)
  maximum time:     1.184 μs (96.23% GC)
  --------------
  samples:          10000
  evals/sample:     998
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Interpolations.ubounds($itp)
BenchmarkTools.Trial:
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     16.777 ns (0.00% GC)
  median time:      18.742 ns (0.00% GC)
  mean time:        22.057 ns (3.95% GC)
  maximum time:     1.865 μs (95.66% GC)
  --------------
  samples:          10000
  evals/sample:     998
  time tolerance:   5.00%
  memory tolerance: 1.00%
```