@timholy 2017-02-13
Hmm, I'm surprised it's different. Is it enough of an improvement that you're fine with it? Or can you see that this has a cost in "real" code? With apologies for my inconsistent statements, we could go back to `@generated` if you think the gap makes a difference.