@rdeits 2017-02-13
My understanding is that the ntuple form creates the closure `i -> size(itp, i)` which contains a reference to non-bitstype `itp` and therefore allocates. The `@generated` form, on the other hand, doesn't create a closure and therefore can avoid allocation entirely. 

Honestly, I've forgotten why I needed `size()` to be faster. I think it came up when I was trying to optimize part of https://github.com/rdeits/regiontrees.jl down to zero allocations, so, for that case, the `@generated` form has an advantage. 

But I'm happy with whatever you prefer. I know there's a general trend to try to avoid `@generated` functions (something about it making certain compiler optimizations harder?).