@timholy 2017-02-14
Does it help if you add `@inline` in front of those definitions?