@rdeits 2017-02-14
No, `@inline` seems to make no difference: 

```
julia> @benchmark size($itp)
BenchmarkTools.Trial:
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     16.954 ns (0.00% GC)
  median time:      17.601 ns (0.00% GC)
  mean time:        19.300 ns (3.52% GC)
  maximum time:     1.065 μs (95.72% GC)
  --------------
  samples:          10000
  evals/sample:     998
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Interpolations.lbounds($itp)
BenchmarkTools.Trial:
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     9.969 ns (0.00% GC)
  median time:      11.243 ns (0.00% GC)
  mean time:        12.644 ns (6.40% GC)
  maximum time:     1.443 μs (97.08% GC)
  --------------
  samples:          10000
  evals/sample:     999
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark Interpolations.ubounds($itp)
BenchmarkTools.Trial:
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     16.784 ns (0.00% GC)
  median time:      17.393 ns (0.00% GC)
  mean time:        18.690 ns (3.32% GC)
  maximum time:     1.088 μs (96.15% GC)
  --------------
  samples:          10000
  evals/sample:     997
  time tolerance:   5.00%
  memory tolerance: 1.00%
```