@timholy 2017-02-14
OK, let's go with this version, and anyone who really needs it can go the `@generated` route.

Thanks, @rdeits!