@rdeits 2017-02-14
Looks like type inference of `size()` is failing in julia v0.4. Is that because it can't infer the return type of the anonymous function `i -> size(itp, i)` ?