@rdeits 2017-02-14
I think *that* might be a sufficiently compelling argument for the `@generated` version, which should be inferred correctly in any julia version.